fix: merge UserInfo entity to User entity
