+++ src/main/java/kr/or/cola/backend/user/UserService.java
@@ -5,8 +5,6 @@ import kr.or.cola.backend.mail.auth.AuthTokenService;
 import kr.or.cola.backend.user.config.auth.dto.OAuthAttributes;
 import kr.or.cola.backend.user.config.auth.dto.SessionUser;
 import kr.or.cola.backend.user.domain.User;
-import kr.or.cola.backend.userinfo.domain.UserInfo;
-import kr.or.cola.backend.userinfo.domain.UserInfoRepository;
 import kr.or.cola.backend.user.domain.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
@@ -26,7 +24,6 @@ import java.util.Collections;
 public class UserService implements OAuth2UserService<OAuth2UserRequest, OAuth2User> {
     private  final UserRepository userRepository;
     private final HttpSession httpSession;
-    private final UserInfoRepository userInfoRepository;
     private final AuthTokenService authTokenService;
 
     @Override
@@ -46,22 +43,29 @@ public class UserService implements OAuth2UserService<OAuth2UserRequest, OAuth2U
         return new DefaultOAuth2User(Collections.singleton(new SimpleGrantedAuthority(user.getRoleKey())), attributes.getAttributes(), attributes.getNameAttributeKey());
     }
 
-    private User saveOrUpdate(OAuthAttributes attributes) {
-        User user = userRepository.findByEmail(attributes.getEmail())
-            .orElse(attributes.toEntity());
-
-        return userRepository.save(user);
-    }
-
     /**
      * 이메일 인증 로직
      * @param token
      */
     public String confirmEmail(String token) {
         AuthToken findAuthToken = authTokenService.findByTokenAndExpirationDateAfterAndExpired(token);
-        UserInfo findUserInfo = userInfoRepository.findByUserId(findAuthToken.getUserId());
+        User user = userRepository.findById(findAuthToken.getUserId())
+            .orElseThrow(() -> new RuntimeException("invalid user id"));
         findAuthToken.useToken();	// 토큰 만료 로직을 구현해주면 된다. ex) expired 값을 true로 변경
-        findUserInfo.emailVerifiedSuccess();	// 유저의 이메일 인증 값 변경 로직을 구현해주면 된다. ex) emailVerified 값을 true로 변경
+        user.emailVerifiedSuccess();	// 유저의 이메일 인증 값 변경 로직을 구현해주면 된다. ex) emailVerified 값을 true로 변경
         return "성공 짝짝";
+    }
+
+    public User findUserByEmail(String email) {
+        return userRepository.findByEmail(email).orElseThrow(() ->
+            new RuntimeException("invalid user")
+        );
+    }
+
+    private User saveOrUpdate(OAuthAttributes attributes) {
+        User user = userRepository.findByEmail(attributes.getEmail())
+            .orElse(attributes.toEntity());
+
+        return userRepository.save(user);
     }
 }
