--- src/main/java/kr/or/cola/backend/userinfo/domain/UserInfoRepository.java
@@ -1,8 +1,0 @@
-package kr.or.cola.backend.userinfo.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserInfoRepository extends JpaRepository<UserInfo, Long> {
-    UserInfo findByUserId(Long userId);
-    boolean existsByUserId(Long userId);
-}
